#6 Update the SEIRModel Class
